+++ src/main/java/frc/robot/poseestimator/helpers/RobotHeadingEstimator.java
@@ -5,7 +5,7 @@ import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.interpolation.TimeInterpolatableBuffer;
 import frc.constants.RobotHeadingEstimatorConstants;
 import frc.robot.poseestimator.PoseEstimatorMath;
-import frc.robot.poseestimator.helpers.RingBuffer.RingBuffer;
+import frc.utils.buffers.RingBuffer.RingBuffer;
 import frc.utils.TimedValue;
 import frc.utils.math.PoseMath;
