+++ src/main/java/frc/utils/buffers/RingBuffer/RingBuffer.java
@@ -1,4 +1,4 @@
-package frc.robot.poseestimator.helpers.RingBuffer;
+package frc.utils.buffers.RingBuffer;
 
 
 import java.util.ArrayList;
